# perf/perf_metrics: Improve parsing logic