perf mertrics test parses the o/p of perf list command to identify
list of metrics that are then passed to perf stat command. The
parsing logic does not work correctly which causes failures as
mentioned:

perf stat -M executing in the branch unit] -C 0 sleep 1'
 [stderr] Cannot find metric or group `executing'
 [stderr] Workload failed: No such file or directory
Command 'perf stat -M executing in the branch unit] -C 0 sleep 1' 
finished with 255 after 0.072842537s

Update the parsing logic to also look for lines which has all
upper case characters.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>