@sacsant 2023-08-08
Before the fix:

 (1/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_M: STARTED
 (1/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_M: FAIL: perf_metric: commands failed are ['perf stat -M executing in the branch unit] -C 0 sleep 1', 'perf stat -M cannot complete because the thread was blocked] -C 0 sleep 1', 'perf stat -M recovering from a TLB miss] -C 0 sleep 1', 'perf stat -M load or store ... (213.83 s)
 (2/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_metric: STARTED
 (2/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_metric: FAIL: perf_metric: commands failed are ['perf stat --metric-only -M executing in the branch unit] -C 0 sleep 1', 'perf stat --metric-only -M cannot complete because the thread was blocked] -C 0 sleep 1', 'perf stat --metric-only -M recovering from a TLB miss] -... (213.98 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 2 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0

After the fix:

 (1/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_M: STARTED
 (1/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_M: PASS (189.83 s)
 (2/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_metric: STARTED
 (2/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_metric: PASS (189.82 s)
RESULTS    : PASS 2 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0